chore(pr) Renamed common-projects.tsx.
