+++ editor/src/components/canvas/canvas-strategies/strategies/common-projects.test-utils.tsx
